code as of 10/21 - finalized SwerveWheel class - coded swerve drive algorithm w/ optional relative steering - made SwerveDriveTestHarness more robust 	- added 3 modes (manual, vector, swerve) for incremental testing 	- added buttons to control modes, relative steering, and gyro reset 	- added more output fields

+++ TestProj/src/testproj/Math708.java
@@ -5,14 +5,22 @@
 package testproj;
 
 /**
- *
+ * Miscellaneous math functions.
  * @author 708
  */
 public class Math708 {
     
+    /**
+     * Used to round a floating point value to
+     * a given decimal place.
+     * @param n The floating point value.
+     * @param dig The decimal place to round to (use negative
+     * numbers for places > 0.
+     * @return 
+     */
     public static double round(double n,int dig)
     {
-        int tens = (int)Math.pow(10,dig);
-        return (double)((int)((n * tens) + .5)) / tens;
+        double tens = Math.pow(10,dig);
+        return ((int)((n * tens) + .5)) / tens;
     }
 }
